+++ pom.xml
@@ -7,7 +7,7 @@
     <artifactId>pupi</artifactId>
     <version>0.0.1</version>
     <name>pupi</name>
-    <packaging>war</packaging>
+    <packaging>jar</packaging>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
