+++ src/main/resources/application.yml
@@ -14,4 +14,4 @@ spring:
       hibernate:
         dialect: org.hibernate.dialect.MySQL5Dialect
 server:
-  port: 8080
+  port: 5000
