+++ src/main/webapp/views/cuidadores/cuidadores-perfil.html
@@ -468,7 +468,7 @@
 <script src="/js/axios.min.js"></script>
 <link rel="stylesheet" href="/css/sweetalert.css" />
 <script src="/js/sweetalert.min.js"></script>
-<script type="text/javascript" src="/views/security/js/login.js"></script>
+<script type="text/javascript" src="/views/security/js/vue-login.js"></script>
 <script src="js/cuidadores-perfil/cuidadores-perfil.js"></script>
 
 <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBxRjrAgwzUG6g4thS0f1osXdML4Npc3cQ&libraries=places"></script>
