+++ src/main/webapp/views/cuidadores/lista-cuidadores.html
@@ -399,7 +399,7 @@
 <!-- Maps -->
 <script type="text/javascript"
         src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBxRjrAgwzUG6g4thS0f1osXdML4Npc3cQ&libraries=places"></script>
-<script type="text/javascript" src="/views/security/js/login.js"></script>
+<script type="text/javascript" src="/views/security/js/vue-login.js"></script>
 <script type="text/javascript" src="/views/cuidadores/js/cuidadores.js"></script>
