+++ src/main/webapp/views/dashboard/dashboard.html
@@ -22,7 +22,8 @@
 
 <!-- Wrapper -->
 <div id="wrapper">
-    <my-dashboard current-view="my-profile"></my-dashboard>
+    <!-- render the currently active component/page here -->
+    <my-dashboard current-view="my-main-dashboard"></my-dashboard>
 </div>
 <!-- Wrapper / End -->
 
@@ -41,8 +42,7 @@
 <script type="text/javascript" src="/assets/scripts/tooltips.min.js"></script>
 <script type="text/javascript" src="/assets/scripts/custom.js"></script>
 
-<link href="/assets/css/plugins/datedropper.css" rel="stylesheet" type="text/css">
-<script src="/assets/scripts/datedropper.js"></script>
+
 <script src="/js/vue.js"></script>
 <script src="/js/axios.min.js"></script>
 <link rel="stylesheet" href="/css/sweetalert.css"/>
@@ -50,20 +50,14 @@
 <script type="text/javascript" src="/views/security/js/vue-login.js"></script>
 <script type="text/javascript" src="/views/dashboard/js/vue-header.js"></script>
 <script type="text/javascript" src="/views/dashboard/js/vue-sidebar.js"></script>
-<!--own-->
-<script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBxRjrAgwzUG6g4thS0f1osXdML4Npc3cQ&libraries=places"></script>
-<script src="js/vue-profile.js"></script>
-
-
-<!--INIT-->
 <script type="text/javascript" src="/views/dashboard/js/vue-dashboard.js"></script>
+<script src="js/vue-main-dashboard.js"></script>
 
 <script>
     let vm = new Vue({
         el: '#wrapper'
     });
 </script>
-
 
 </body>
 </html>
