+++ src/main/webapp/views/dashboard/js/vue-main-dashboard.js
@@ -1,0 +1,54 @@
+let myMainDashboard = Vue.component('my-main-dashboard', {
+    template: `
+<div>
+        <!-- Titlebar -->
+		<div id="titlebar">
+			<div class="row">
+				<div class="col-md-12">
+					<h2>Resumen</h2> 
+				</div>
+			</div>
+		</div>
+
+		<!-- Content -->
+		<div class="row">
+
+			<!-- Item -->
+			<div class="col-lg-3 col-md-6">
+				<div class="dashboard-stat color-1">
+					<div class="dashboard-stat-content"><h4>6</h4> <span>Active Listings</span></div>
+					<div class="dashboard-stat-icon"><i class="im im-icon-Map2"></i></div>
+				</div>
+			</div>
+
+			<!-- Item -->
+			<div class="col-lg-3 col-md-6">
+				<div class="dashboard-stat color-2">
+					<div class="dashboard-stat-content"><h4>726</h4> <span>Total Views</span></div>
+					<div class="dashboard-stat-icon"><i class="im im-icon-Line-Chart"></i></div>
+				</div>
+			</div>
+
+			
+			<!-- Item -->
+			<div class="col-lg-3 col-md-6">
+				<div class="dashboard-stat color-3">
+					<div class="dashboard-stat-content"><h4>95</h4> <span>Total Reviews</span></div>
+					<div class="dashboard-stat-icon"><i class="im im-icon-Add-UserStar"></i></div>
+				</div>
+			</div>
+
+			<!-- Item -->
+			<div class="col-lg-3 col-md-6">
+				<div class="dashboard-stat color-4">
+					<div class="dashboard-stat-content"><h4>126</h4> <span>Times Bookmarked</span></div>
+					<div class="dashboard-stat-icon"><i class="im im-icon-Heart"></i></div>
+				</div>
+			</div>
+		</div>
+
+
+</div>
+`
+});
+
