+++ src/main/webapp/views/dashboard/js/vue-sidebar.js
@@ -1,0 +1,22 @@
+let mySidebar = Vue.component('my-sidebar', {
+    template: `
+        <div class="dashboard-nav">
+            <div class="dashboard-nav-inner">
+
+                <ul data-submenu-title="Principal">
+                    <li class="active"><a href="/views/dashboard/dashboard.html"><i class="sl sl-icon-settings"></i> Dashboard</a></li>
+                </ul>
+
+                <ul data-submenu-title="Secciones">
+                    <li><a href="/views/perros/perros.html"><i class="im im-icon-Dog"></i> Mis Perros</a></li>
+                </ul>
+
+                <ul data-submenu-title="Cuenta">
+                    <li><a href="/views/due%C3%B1o/perfil.html"><i class="sl sl-icon-user"></i> Mi Perfil</a></li>
+                </ul>
+
+            </div>
+        </div>
+        <!-- Navigation / End -->
+`
+});
