+++ src/main/webapp/views/index/css/index.css
@@ -1,10 +1,10 @@
 /*jose*/
 .left-side {
-    width: 65%;
+    width: 70%;
 }
 
 .right-side {
-    width: 35%;
+    width: 30%;
 }
 
 /*jose*/
