+++ src/main/webapp/views/index/index.html
@@ -22,7 +22,7 @@
 <body >
 
 <!-- Wrapper -->
-<div id="appVue">
+<div id="wrapper">
 
 
         <!-- Header Container
@@ -364,7 +364,7 @@
 <!-- Maps -->
 <script type="text/javascript"
         src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBxRjrAgwzUG6g4thS0f1osXdML4Npc3cQ&libraries=places"></script>
-<script type="text/javascript" src="/views/security/js/login.js"></script>
+<script type="text/javascript" src="/views/security/js/vue-login.js"></script>
 <script type="text/javascript" src="/views/index/js/index.js"></script>
