+++ src/main/webapp/views/reserva/reserva.html
@@ -234,7 +234,7 @@
 <script src="/js/vue.js"></script>
 <link rel="stylesheet" href="/css/sweetalert.css"/>
 <script src="/js/sweetalert.min.js"></script>
-<script type="text/javascript" src="/views/security/js/login.js"></script>
+<script type="text/javascript" src="/views/security/js/vue-login.js"></script>
 <script src="js/reserva.js"></script>
 
 <style>
